requirements.txt
tensorflow==2.21.0
numpy
matplotlib
pillow
gradio
scikit-learn
